Transição de métodos da camada de serviço para camada de repositório
