+++ .run/gprojuridico.run.xml
@@ -3,7 +3,7 @@
     <module name="gprojuridico" />
     <extension name="coverage">
       <pattern>
-        <option name="PATTERN" value="com.uniprojecao.fabrica.gprojuridico.repository.*" />
+        <option name="PATTERN" value="com.uniprojecao.fabrica.gprojuridico.repositories.*" />
         <option name="ENABLED" value="true" />
       </pattern>
       <pattern>
@@ -11,7 +11,7 @@
         <option name="ENABLED" value="true" />
       </pattern>
     </extension>
-    <option name="PACKAGE_NAME" value="com.uniprojecao.fabrica.gprojuridico.repository" />
+    <option name="PACKAGE_NAME" value="com.uniprojecao.fabrica.gprojuridico.repositories" />
     <option name="MAIN_CLASS_NAME" value="" />
     <option name="METHOD_NAME" value="" />
     <option name="TEST_OBJECT" value="directory" />
