+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/services/AtendimentoService.java
@@ -1,7 +1,7 @@
 package com.uniprojecao.fabrica.gprojuridico.services;
 
 import com.uniprojecao.fabrica.gprojuridico.models.atendimento.Atendimento;
-import com.uniprojecao.fabrica.gprojuridico.repository.BaseRepository;
+import com.uniprojecao.fabrica.gprojuridico.repositories.FirestoreRepository;
 import org.springframework.stereotype.Service;
 
 import static com.uniprojecao.fabrica.gprojuridico.services.IdService.defineId;
@@ -12,7 +12,7 @@ public class AtendimentoService {
 
     public Atendimento insert(Atendimento data) {
         data = (Atendimento) defineId(data, ATENDIMENTOS_COLLECTION, "ATE");
-        BaseRepository.insert(ATENDIMENTOS_COLLECTION, data.getId(), data);
+        FirestoreRepository.insert(ATENDIMENTOS_COLLECTION, data.getId(), data);
         return data;
     }
 }
