+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/services/MedidaJuridicaService.java
@@ -1,7 +1,7 @@
 package com.uniprojecao.fabrica.gprojuridico.services;
 
 import com.uniprojecao.fabrica.gprojuridico.models.MedidaJuridicaModel;
-import com.uniprojecao.fabrica.gprojuridico.repository.BaseRepository;
+import com.uniprojecao.fabrica.gprojuridico.repositories.FirestoreRepository;
 import org.springframework.stereotype.Service;
 
 import static com.uniprojecao.fabrica.gprojuridico.services.IdService.defineId;
@@ -12,7 +12,7 @@ public class MedidaJuridicaService {
 
     public MedidaJuridicaModel insert(MedidaJuridicaModel model) {
         model = (MedidaJuridicaModel) defineId(model, MEDIDAS_JURIDICAS_COLLECTION, "MEDJUR");
-        BaseRepository.insert(MEDIDAS_JURIDICAS_COLLECTION, model.getId(), model);
+        FirestoreRepository.insert(MEDIDAS_JURIDICAS_COLLECTION, model.getId(), model);
         return model;
     }
 }
