+++ src/main/java/com/uniprojecao/fabrica/gprojuridico/services/UsuarioService.java
@@ -1,11 +1,9 @@
 package com.uniprojecao.fabrica.gprojuridico.services;
 
 import at.favre.lib.crypto.bcrypt.BCrypt;
-import com.uniprojecao.fabrica.gprojuridico.models.autocomplete.UsuarioAutocomplete;
 import com.uniprojecao.fabrica.gprojuridico.models.usuario.Estagiario;
 import com.uniprojecao.fabrica.gprojuridico.models.usuario.Usuario;
-import com.uniprojecao.fabrica.gprojuridico.repository.BaseRepository;
-import com.uniprojecao.fabrica.gprojuridico.repository.UsuarioRepository;
+import com.uniprojecao.fabrica.gprojuridico.repositories.FirestoreRepository;
 import com.uniprojecao.fabrica.gprojuridico.services.exceptions.UserAlreadyCreatedException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -14,53 +12,15 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
-import static com.uniprojecao.fabrica.gprojuridico.services.QueryFilterService.getFilter;
 import static com.uniprojecao.fabrica.gprojuridico.utils.Constants.USUARIOS_COLLECTION;
 import static com.uniprojecao.fabrica.gprojuridico.utils.Utils.filterValidKeys;
-import static java.lang.Integer.parseInt;
 
-/**
- * Baseado no padrão arquitetural camada de serviço, este serviço tem por
- * objetivo apresentar a implementação de regras e lógica de negócio. <p>
- *
- * Este serviço em si trata da interação com o modelo de domínio {@link Usuario},
- * onde as regras de negócios pertinentes a este domínio estão listadas logo
- * abaixo: <p>
- *
- * <h2>Regras de Negócio</h2>
- *
- * <h3>Verificar Usuário</h3>
- * Ao inserir um usuário, deverá ser verificado se este existe no banco de dados.
- * Caso exista, uma exceção deverá ser lançada; senão o fluxo segue-se normalmente. <p>
- *
- * <h3>Gerar Senha Temporária</h3>
- * Caso o DTO recebido não apresente a senha preenchida, uma senha aleatória deverá
- * ser gerada. Além disso, o usuário deve receber, por e-mail, suas credenciais de
- * acesso, contendo: nome de usuário, e-mail e senha temporária. <p>
- *
- * <h3>Gerar Senha Definitiva</h3>
- * Caso seja a 1ª vez que o usuário efetue login no sistema e possua senha temporária
- * ativa, este deverá trocar a senha temporária por uma definitiva, que será informada
- * pelo próprio usuário. <p>
- *
- * <h3>Criptografar Senha</h3>
- * Após o sistema receber o usuário com a senha (definitiva) preenchida, o sistema
- * deverá criptografar a senha para o correto armazenamento desta em algum banco de
- * dados. <p>
- *
- * @author Jonatas Mateus
- * @see Usuario
- * @since 1.0
- */
 @Service
 public class UsuarioService implements UserDetailsService {
 
-    public UsuarioRepository repository = new UsuarioRepository();
-
-    private void checkIfExists(Usuario data, String id) {
+    private void checkIfExists(Usuario data, String id) throws Exception {
         var result = findById(id);
         if (result != null) {
             String userEmailFound = result.getEmail();
@@ -92,35 +52,35 @@ public class UsuarioService implements UserDetailsService {
         return BCrypt.withDefaults().hashToString(12, rawPassword.toCharArray());
     }
 
-    public List<UsuarioAutocomplete> findAllMin(String limit, String field, String filter, String value) {
-        return repository.findAllMin(parseInt(limit), getFilter(field, filter, value));
+    public Usuario findById(String id) throws Exception {
+        return (Usuario) FirestoreRepository.getDocumentById(USUARIOS_COLLECTION, id);
     }
 
-    public Usuario findById(String id) {
-        return repository.findById(id);
-    }
-
-    public Usuario findMe() {
+    public Usuario findMe() throws Exception {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
         Usuario usuario = (Usuario) authentication.getPrincipal();
         String id = usuario.getId();
         return findById(id);
     }
 
-    public Usuario insert(Usuario data) {
+    public Usuario insert(Usuario data) throws Exception {
         defineId(data);
         String id = data.getId();
 
         checkIfExists(data, id);
         encryptPassword(data);
 
-        BaseRepository.insert(USUARIOS_COLLECTION, id, data);
+        FirestoreRepository.insert(USUARIOS_COLLECTION, id, data);
         return data;
     }
 
     @Override
     public UserDetails loadUserByUsername(String username) {
-        return repository.findById(username);
+        try {
+            return (UserDetails) FirestoreRepository.getDocumentById(USUARIOS_COLLECTION, username);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
     }
 
     public void update(String id, Map<String, Object> data, String clazz) {
@@ -145,9 +105,9 @@ public class UsuarioService implements UserDetailsService {
             newData.put(senhaKey, encryptedPassword);
 
             // Salva os dados recebidos e com a senha criptografada
-            BaseRepository.update(USUARIOS_COLLECTION, id, newData);
+            FirestoreRepository.update(USUARIOS_COLLECTION, id, newData);
         } else {
-            BaseRepository.update(USUARIOS_COLLECTION, id, filteredData);
+            FirestoreRepository.update(USUARIOS_COLLECTION, id, filteredData);
         }
     }
 }
